requirements.txt
adrf==0.1.4
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
dj-database-url==2.1.0
Django==4.1.6
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
gotrue==2.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pluggy==1.3.0
postgrest==0.15.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyotp==2.9.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
realtime==1.0.2
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
